Richard S_E2E/pyproject.toml
[tool.poetry]
name = "chrome-gpt"
version = "0.1.0"
description = "An AutoGPT agent that uses Chrome interactively on your desktop"
authors = ["Richard He <he.yucheng@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0" 
langchain = ">=0.0.157"
pytest = "^7.3.1"
mypy = "^1.2.0"
black = "^23.3.0"
ruff = "^0.0.263"
faiss-cpu = "^1.7.4"
click = "^8.1.3"
langchain_google_genai = "^1.0.6"
tiktoken = "^0.3.3"
validators = "^0.20.0"
pexpect = "^4.8.0"
unidecode = "1.3.6"
bs4 = "^0.0.1"
selenium = "^4.13.0"

[tool.ruff]
select = [
  "E",  # pycodestyle
  "F",  # pyflakes
  "I",  # isort
]

[tool.mypy]
ignore_missing_imports = "True"
disallow_untyped_defs = "True"
exclude = ["notebooks"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"